# PERF: grouped_reduce