```
from asv_bench.benchmarks.groupby import *
self = GroupByMethods()
self.setup("int", "std", "transformation", 10)

%timeit self.time_dtype_as_field('int', 'std', 'transformation', 10)
325 µs ± 9.84 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
361 µs ± 2.05 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
```